remove compiler warnings

+++ skeletons/per_support.c
@@ -13,9 +13,9 @@ per_data_string(asn_per_data_t *pd) {
 	static int n;
 	n = (n+1) % 2;
 	snprintf(buf[n], sizeof(buf),
-		"{m=%ld span %+d[%d..%d] (%d)}",
+		"{m=%ld span %+ld[%d..%d] (%d)}",
 		(long)pd->moved,
-		(((int)pd->buffer) & 0xf),
+		(((long)pd->buffer) & 0xf),
 		(int)pd->nboff, (int)pd->nbits,
 		(int)(pd->nbits - pd->nboff));
 	return buf[n];
@@ -104,10 +104,10 @@ per_get_few_bits(asn_per_data_t *pd, int nbits) {
 
 	accum &= (((uint32_t)1 << nbits) - 1);
 
-	ASN_DEBUG("  [PER got %2d<=%2d bits => span %d %+d[%d..%d]:%02x (%d) => 0x%x]",
+	ASN_DEBUG("  [PER got %2d<=%2d bits => span %d %+ld[%d..%d]:%02x (%d) => 0x%x]",
 		(int)nbits, (int)nleft,
 		(int)pd->moved,
-		(((int)pd->buffer) & 0xf),
+		(((long)pd->buffer) & 0xf),
 		(int)pd->nboff, (int)pd->nbits,
 		pd->buffer[0],
 		(int)(pd->nbits - pd->nboff),
